+++ src/utils/architect/tradeMachine/cbaConstants.js
@@ -1,0 +1,10 @@
+export const CBA_THRESHOLDS = {
+  FIRST_APRON: 172_295_000,
+  SECOND_APRON: 182_794_000,
+  MIN_SALARY: 1_119_563,
+  MAX_CASH: 7_100_000,
+  TRADE_BUFFER: 100_000,
+  STEPIEN_YEARS: 7,
+};
+
+export const MAX_FUTURE_PICK_YEARS = 7;
