+++ src/utils/architect/tradeMachine/tradeUtils.js
@@ -1,0 +1,27 @@
+export const isExpired = (expiryDate) => {
+  if (!expiryDate) return false;
+  const expiry = new Date(expiryDate);
+  const today = new Date();
+  today.setHours(0, 0, 0, 0); // Normalize to midnight
+  return expiry <= today;
+};
+
+export const formatDate = (dateStr) => {
+  if (!dateStr) return 'N/A';
+  return new Date(dateStr).toLocaleDateString('en-US', {
+    month: 'short',
+    day: 'numeric',
+    year: 'numeric',
+  });
+};
+
+export const formatSalary = (amount) => `$${(amount || 0).toLocaleString()}`;
+
+export const isMeaningfulProtection = (protection) => {
+  if (!protection) return false;
+  return (
+    /top\s*[1-9]\d*/i.test(protection) ||
+    /lottery/i.test(protection) ||
+    /1-14/i.test(protection)
+  );
+};
