+++ src/utils/architect/tradeMachine/tradeValidator.js
@@ -37,894 +37,37 @@ import {
   violates2MonthAggregation,
 } from '@/utils/architect/timingUtils.js';
 import { collectConsentViolations } from '@/utils/architect/consentUtils.js';
-import { getReacqBlock } from '@/utils/architect/reacqUtils.js';
+import { isMeaningfulProtection } from '@/utils/architect/tradeMachine/tradeUtils.js';
+import debug from '@/utils/architect/tradeMachine/tradeDebug.js';
+
 import {
-  getSeasonalCashCaps,
-  computeSeasonCashLedger,
-} from '@/utils/architect/cashUtils.js';
-
-// ===== DEBUGGER =====
-const debug = {
-  enabled: false,
-  logs: [],
-  records: [],
-
-  /**
-   * Append a message to the debug log. The message is printed to the
-   * console for immediate visibility and stored so it can be flushed to
-   * a text file later on.
-   */
-  write(msg = '', meta = {}) {
-    if (!this.enabled) return;
-    if (msg !== '' || this.logs.length > 0) {
-      this.logs.push(msg);
-      this.records.push({
-        msg,
-        time: new Date(),
-        team: meta.team,
-        rule: meta.rule,
-        salary: meta.salary || false,
-      });
-    }
-    console.log(msg);
-  },
-
-  log(msg, meta = {}) {
-    this.write(msg, meta);
-  },
-
-  logTrade(team) {
-    this.log('', { team: team.team.teamName });
-    this.log(`=== ${team.team.teamName} ===`, { team: team.team.teamName });
-    this.log(`Current Salary: ${formatCurrency(team.team.totalSalary)}`, {
-      team: team.team.teamName,
-    });
-    this.log(
-      `Status: ${getApronStatus(team.team.totalSalary, team.context.capSettings)}`,
-      { team: team.team.teamName }
-    );
-  },
-
-  logSalaries(team) {
-    this.log('Outgoing:', { team: team.team.teamName, salary: true });
-    (team.sends || []).forEach((p) => {
-      const salary =
-        p.contract_clean?.salaries_by_year?.[team.context.yearKey]?.salary || 0;
-      this.log(`  - ${p.name}: ${formatCurrency(salary)}`, {
-        team: team.team.teamName,
-        salary: true,
-      });
-    });
-
-    this.log('Incoming:', { team: team.team.teamName, salary: true });
-    (team.incomingPlayers || []).forEach((p) => {
-      const salary =
-        p.contract_clean?.salaries_by_year?.[team.context.yearKey]?.salary || 0;
-      this.log(`  - ${p.name}: ${formatCurrency(salary)}`, {
-        team: team.team.teamName,
-        salary: true,
-      });
-    });
-
-    this.log(
-      `Totals: OUT ${formatCurrency(team.salaryOut)} | IN ${formatCurrency(team.salaryIn)}`,
-      { team: team.team.teamName, salary: true }
-    );
-  },
-
-  logSecondApron(team, violations) {
-    if (!team.overSecondApron && !team.willBeOverSecond) return;
-
-    this.log('Second Apron Rules:', {
-      team: team.team.teamName,
-      rule: 'secondApron',
-    });
-    this.log(
-      `Trade Type: ${team.sends.length}-for-${team.incomingPlayers.length}`,
-      {
-        team: team.team.teamName,
-        rule: 'secondApron',
-      }
-    );
-
-    if (team.sends.length === 1) {
-      const outgoing =
-        team.sends[0].contract_clean?.salaries_by_year?.[team.context.yearKey]
-          ?.salary || 0;
-      this.log(`1-to-Many max incoming: ${formatCurrency(outgoing)}`, {
-        team: team.team.teamName,
-        rule: 'secondApron',
-      });
-    } else {
-      this.log('Many-to-Many: pair incoming with outgoing', {
-        team: team.team.teamName,
-        rule: 'secondApron',
-      });
-    }
-
-    if (violations.length) {
-      this.log('Violations:', {
-        team: team.team.teamName,
-        rule: 'secondApron',
-      });
-      violations.forEach((v) =>
-        this.log(`  - ${v}`, { team: team.team.teamName, rule: 'secondApron' })
-      );
-    } else {
-      this.log('All rules satisfied', {
-        team: team.team.teamName,
-        rule: 'secondApron',
-      });
-    }
-
-    // Spacer between team logs for readability
-    this.log('', { team: team.team.teamName, rule: 'secondApron' });
-  },
-
-  async flush(file = 'trade-debug.txt') {
-    if (typeof process !== 'undefined' && process.versions?.node) {
-      const fs = await import('fs');
-      await fs.promises.writeFile(file, this.logs.join('\n'), 'utf8');
-      console.log(`\nDebug log written to ${file}`);
-    }
-  },
-
-  flushToUI({ showSalary = true, team = null, rule = null } = {}) {
-    return this.records
-      .filter((r) => {
-        if (!showSalary && r.salary) return false;
-        if (team && r.team !== team) return false;
-        if (rule && r.rule !== rule) return false;
-        return true;
-      })
-      .map((r) => `${r.time.toLocaleTimeString()} - ${r.msg}`);
-  },
-};
-
-// ===== UTILITIES =====
-const isExpired = (expiryDate) => {
-  if (!expiryDate) return false;
-  const expiry = new Date(expiryDate);
-  const today = new Date();
-  today.setHours(0, 0, 0, 0); // Normalize to midnight
-  return expiry <= today;
-};
-
-const formatDate = (dateStr) => {
-  if (!dateStr) return 'N/A';
-  return new Date(dateStr).toLocaleDateString('en-US', {
-    month: 'short',
-    day: 'numeric',
-    year: 'numeric',
-  });
-};
-
-const formatSalary = (amount) => `$${(amount || 0).toLocaleString()}`;
-
-const isMeaningfulProtection = (protection) => {
-  if (!protection) return false;
-  return (
-    /top\s*[1-9]\d*/i.test(protection) ||
-    /lottery/i.test(protection) ||
-    /1-14/i.test(protection)
-  );
-};
-
-// ===== CBA CONSTANTS =====
-const CBA_THRESHOLDS = {
-  FIRST_APRON: 172_295_000,
-  SECOND_APRON: 182_794_000,
-  MIN_SALARY: 1_119_563,
-  MAX_CASH: 7_100_000,
-  TRADE_BUFFER: 100_000,
-  STEPIEN_YEARS: 7,
-};
-
-const MAX_FUTURE_PICK_YEARS = 7;
+  CBA_THRESHOLDS,
+  MAX_FUTURE_PICK_YEARS,
+} from '@/utils/architect/tradeMachine/cbaConstants.js';
+import { enforceSecondApronHandcuffs } from '@/utils/architect/tradeMachine/rules/enforceSecondApronHandcuffs.js';
+import { enforceRosterWindow } from '@/utils/architect/tradeMachine/rules/enforceRosterWindow.js';
+import { enforceConsent } from '@/utils/architect/tradeMachine/rules/enforceConsent.js';
+import { enforceEligibility } from '@/utils/architect/tradeMachine/rules/enforceEligibility.js';
+import { enforceTiming } from '@/utils/architect/tradeMachine/rules/enforceTiming.js';
+import { validateTradeExceptions } from '@/utils/architect/tradeMachine/validators/validateTradeExceptions.js';
+import { validateFaExceptionUsage } from '@/utils/architect/tradeMachine/validators/validateFaExceptionUsage.js';
+import { validateDraftPicks } from '@/utils/architect/tradeMachine/validators/validateDraftPicks.js';
+import { validateCash } from '@/utils/architect/tradeMachine/validators/validateCash.js';
+import { validateSignAndTrade } from '@/utils/architect/tradeMachine/validators/validateSignAndTrade.js';
+import { validateBYC } from '@/utils/architect/tradeMachine/validators/validateBYC.js';
+import { validateSecondApronRules } from '@/utils/architect/tradeMachine/validators/validateSecondApronRules.js';
+import { validateAllNewRules } from '@/utils/architect/tradeMachine/validators/validateAllNewRules.js';
+import { getMatchingValue, computeMatchingValues } from '@/utils/architect/tradeMachine/matchingValues.js';
 
 // ===== SECOND APRON HANDCUFFS =====
-export function enforceSecondApronHandcuffs(teamCtx, tradeCtx = {}) {
-  const violations = [];
-  if (!teamCtx?.postTradeStatus?.isAtOrAboveSecondApron) return violations;
-
-  const outgoing = teamCtx.outgoingPlayers || [];
-  const incoming = teamCtx.incomingPlayers || [];
-
-  const season = teamCtx.context?.yearKey;
-  const usedTPEIds = new Set(
-    incoming.filter((p) => p.acquiredViaTPE && p.tpeId).map((p) => p.tpeId)
-  );
-  if (season != null) {
-    let priorAdded = false;
-    const addPrior = () => {
-      if (!priorAdded) {
-        violations.push('Second apron: prior-year TPEs cannot be used.');
-        priorAdded = true;
-      }
-    };
-    if (usedTPEIds.size) {
-      (teamCtx.tradeExceptions || []).forEach((tpe) => {
-        if (usedTPEIds.has(tpe.id) && !isCurrentSeasonTPE(tpe, season)) {
-          addPrior();
-        }
-      });
-    }
-    if (
-      Array.isArray(teamCtx.appliedTPEs) &&
-      teamCtx.appliedTPEs.some((tpe) => !isCurrentSeasonTPE(tpe, season))
-    ) {
-      addPrior();
-    }
-  }
-
-  if (outgoing.length > 1 && incoming.length <= 1) {
-    violations.push('Second apron teams cannot aggregate salaries');
-  }
-
-  if (teamCtx.cashSent > 0 || teamCtx.cashReceived > 0) {
-    violations.push('Second apron team cannot include cash in trades');
-  }
-
-  if ((teamCtx.salaryIn || 0) > (teamCtx.salaryOut || 0)) {
-    violations.push('Second apron team cannot receive more salary than sent');
-  }
-
-  return violations;
-}
-
-export function enforceRosterWindow(
-  teamCtx,
-  tradeCtx = {},
-  { warn = () => {}, reject = () => {} } = {}
-) {
-  const check = passesRosterWindow(teamCtx.postTradeTeam, {
-    require14to15: !tradeCtx?.graceMode,
-  });
-  const stdMsg = check.reasons.find((r) => r.startsWith('Standard'));
-  const twoWayMsg = check.reasons.find((r) => r.startsWith('Two-way'));
-  if (stdMsg) {
-    if (validationFlags.rosterEnforcement === 'error') reject(stdMsg);
-    if (validationFlags.rosterEnforcement === 'warn') warn(stdMsg);
-  }
-  if (twoWayMsg) {
-    if (validationFlags.twoWayRoster === 'error') reject(twoWayMsg);
-    if (validationFlags.twoWayRoster === 'warn') warn(twoWayMsg);
-  }
-  return check;
-}
-
-export function enforceConsent(
-  teamCtx,
-  tradeCtx = {},
-  { warn = () => {}, reject = () => {} } = {}
-) {
-  const enforcement = validationFlags.consent;
-  const violations = [];
-  const consentMap = tradeCtx.consent || {};
-  const seen = new Set();
-  (teamCtx.outgoingPlayers || []).forEach((p) => {
-    const destId =
-      p.tradeTo ??
-      p.toTeamId ??
-      Object.keys(tradeCtx.teamNames || {}).find((id) => id !== teamCtx.teamId);
-    const consent = consentMap[p.id] || {};
-    const msgs = collectConsentViolations(p, destId, consent, { reject });
-    msgs.forEach((msg) => {
-      if (!seen.has(msg)) {
-        violations.push(msg);
-        seen.add(msg);
-        if (enforcement === 'warn') warn(msg);
-      }
-    });
-  });
-  return violations;
-}
-
-export function enforceEligibility(
-  teamCtx,
-  ctx = {},
-  { warn = () => {}, reject = () => {} } = {}
-) {
-  const enforcement = validationFlags.reAcquisition;
-  const violations = [];
-  const nowDate = ctx.now
-    ? new Date(ctx.now)
-    : ctx.asOfDate
-      ? new Date(ctx.asOfDate)
-      : new Date();
-  (teamCtx.incomingPlayers || []).forEach((p) => {
-    let block = getReacqBlock(p, teamCtx.teamId, nowDate);
-    if (
-      !block.blocked &&
-      typeof ctx.wasTradedAwayWithinOneYear === 'function' &&
-      ctx.wasTradedAwayWithinOneYear(p.id, teamCtx.teamId)
-    ) {
-      const until = new Date(nowDate);
-      until.setFullYear(until.getFullYear() + 1);
-      block = { blocked: true, until };
-    }
-    if (block.blocked) {
-      const dateStr = block.until.toISOString().slice(0, 10);
-      const msg = `Re-acquisition bar: ${teamCtx.teamName} cannot reacquire ${p.name} until ${dateStr}`;
-      violations.push(msg);
-      reject(msg);
-      if (enforcement === 'warn') warn(msg);
-    }
-  });
-  return violations;
-}
-
-export function enforceTiming(
-  teamCtx,
-  tradeCtx = {},
-  { warn = () => {}, reject = () => {} } = {}
-) {
-  const enforcement = validationFlags.timingEnforcement;
-  const tradeDate = new Date(tradeCtx.tradeDate || Date.now());
-
-  if (isWithinMoratorium(tradeDate, validationFlags.moratorium)) {
-    const msg = 'Trade moratorium in effect';
-    if (enforcement === 'error') reject(msg);
-    if (enforcement === 'warn') warn(msg);
-  }
-
-  const outgoing = teamCtx.outgoingPlayers || [];
-  const aggregated = outgoing.length > 1;
-  outgoing.forEach((p) => {
-    const name = p.name || 'Player';
-
-    if (p.eligibleTradeDate && tradeDate < new Date(p.eligibleTradeDate)) {
-      const msg = `${name} not trade-eligible until ${p.eligibleTradeDate}`;
-      if (enforcement === 'error') reject(msg);
-      if (enforcement === 'warn') warn(msg);
-    } else if (p.jan15Eligible === false) {
-      const jan15 = new Date(Date.UTC(tradeDate.getUTCFullYear(), 0, 15));
-      if (tradeDate < jan15) {
-        const msg = `${name} not trade-eligible until Jan 15`;
-        if (enforcement === 'error') reject(msg);
-        if (enforcement === 'warn') warn(msg);
-      }
-    }
-
-    if (violates30Day(p, tradeDate)) {
-      const msg = `${name} signed within last 30 days`;
-      if (enforcement === 'error') reject(msg);
-      if (enforcement === 'warn') warn(msg);
-    }
-
-    if (aggregated) {
-      const violates = p.eligibleAggregationDate
-        ? tradeDate < new Date(p.eligibleAggregationDate)
-        : violates2MonthAggregation(p, tradeDate);
-      if (violates) {
-        const msg = `${name} cannot be aggregated in trade yet`;
-        if (enforcement === 'error') reject(msg);
-        if (enforcement === 'warn') warn(msg);
-      }
-    }
-  });
-}
 
 // ===== TRADE EXCEPTION VALIDATION =====
-export function validateTradeExceptions(team) {
-  const violations = [];
-  let addedGeneric = false;
-  if (team.tradeExceptions?.some((e) => e.isBeingUsed)) {
-    team.tradeExceptions
-      .filter((e) => e.isBeingUsed)
-      .forEach((e) =>
-        violations.push(
-          `TPE ${e.id} is already being processed in another transaction`
-        )
-      );
-  }
-  const usedTPEs = new Set();
-  const yearKey = team?.context?.yearKey ?? new Date().getFullYear();
-  const onDate = team.context?.tradeDate || new Date().toISOString();
-
-  const addBlock = () => {
-    if (!addedGeneric) {
-      violations.push(SECOND_APRON_TPE_BLOCK);
-      addedGeneric = true;
-    }
-  };
-
-  team.incomingPlayers.forEach((player) => {
-    if (!player.acquiredViaTPE) return;
-
-    const tpe = team.tradeExceptions.find((e) => e.id === player.tpeId);
-    if (!tpe) {
-      violations.push(`No valid TPE found for ${player.name}`);
-      return;
-    }
-    if (usedTPEs.has(tpe.id)) {
-      violations.push(
-        `TPE ${tpe.id} is being used multiple times in this trade`
-      );
-      return;
-    }
-    if (team.outgoingPlayers?.some((p) => p.toTeamId === player.fromTeamId)) {
-      violations.push('Cannot aggregate trade exception with outgoing salary');
-      return;
-    }
-    if (team.postTradeStatus?.isAtOrAboveSecondApron) {
-      addBlock();
-      return;
-    }
-    if (!canUseTPE(team, tpe, { currentSeason: yearKey, onDate })) {
-      if (isExpiredTPE(tpe, onDate)) {
-        violations.push(`Trade exception ${tpe.id} is expired`);
-      } else {
-        violations.push(`Cannot use trade exception ${tpe.id}`);
-      }
-      return;
-    }
-    const incoming = getSalaryForYear(player, yearKey);
-    if (incoming > tpe.amount) {
-      violations.push(`Trade exception ${tpe.id} is too small`);
-      return;
-    }
-    usedTPEs.add(tpe.id);
-    tpe.remaining = tpe.amount - incoming;
-    tpe.isUsed = tpe.remaining === 0;
-  });
-
-  if (
-    team.postTradeStatus?.isAtOrAboveSecondApron &&
-    Array.isArray(team.appliedTPEs)
-  ) {
-    if (team.appliedTPEs.length) {
-      addBlock();
-    }
-  }
-
-  return violations;
-}
-
-export function validateFaExceptionUsage(team, flags = validationFlags) {
-  const violations = [];
-  let addedGeneric = false;
-  const { faExceptionTrade = 'on', faExceptionAutoSelect = true } = flags;
-  if (faExceptionTrade === 'off') {
-    if (team.incomingPlayers.some((p) => p.absorptionMode === 'FA_EXCEPTION')) {
-      violations.push('FA Exception usage disabled.');
-    }
-    return violations;
-  }
-
-  const ctx = {
-    teamSeasonState: team.team || {},
-    teamTotalSalary: team.teamTotalSalary || 0,
-    context: team.context || {},
-  };
-  const yearKey = ctx.context.yearKey;
-  const buckets = getTeamFaExceptionBuckets(ctx.teamSeasonState);
-  const faPlayers = (team.incomingPlayers || []).filter(
-    (p) =>
-      p.absorptionMode === 'FA_EXCEPTION' ||
-      (faExceptionAutoSelect && !p.absorptionMode)
-  );
-
-  if (
-    faPlayers.length &&
-    ctx.teamTotalSalary >= ctx.context.capSettings?.secondApron
-  ) {
-    violations.push(
-      'Second Apron — trade exceptions are banned for second-apron teams'
-    );
-    if (!addedGeneric) {
-      violations.push(SECOND_APRON_TPE_BLOCK);
-      addedGeneric = true;
-    }
-  }
-
-  faPlayers.forEach((player) => {
-    let mode = player.absorptionMode;
-    let bucketType = player.bucketType;
-    const salary = player.matchIncoming ?? getSalaryForYear(player, yearKey);
-
-    if (!mode && faExceptionAutoSelect) {
-      const bucket = buckets.find(
-        (b) =>
-          isFaExceptionEligibleType(b.type, flags) &&
-          canUseFaException(ctx, b.type) &&
-          b.remaining >= salary
-      );
-      if (bucket) {
-        mode = 'FA_EXCEPTION';
-        bucketType = bucket.type;
-        player.absorptionMode = 'FA_EXCEPTION';
-        player.bucketType = bucketType;
-      }
-    }
-
-    if (mode !== 'FA_EXCEPTION') return;
-
-    if (Array.isArray(bucketType)) {
-      violations.push(
-        'Cannot combine FA Exception with outgoing salary for the same player.'
-      );
-      return;
-    }
-
-    if (ctx.teamTotalSalary >= ctx.context.capSettings?.secondApron) {
-      if (
-        !violations.includes(
-          'FA Exception unavailable above/beyond Second Apron.'
-        )
-      ) {
-        violations.push('FA Exception unavailable above/beyond Second Apron.');
-      }
-      return;
-    }
-    if (ctx.teamTotalSalary >= ctx.context.capSettings?.firstApron) {
-      violations.push('FA Exception unavailable above/beyond First Apron.');
-      return;
-    }
-
-    const bucket = buckets.find((b) => b.type === bucketType);
-    if (!bucket || !isFaExceptionEligibleType(bucketType, flags)) {
-      violations.push(`FA Exception ${bucketType} not available`);
-      return;
-    }
-    if (team.outgoingPlayers?.some((p) => p.toTeamId === player.fromTeamId)) {
-      violations.push(
-        'Cannot combine FA Exception with outgoing salary for the same player.'
-      );
-      return;
-    }
-    if (salary > bucket.remaining) {
-      violations.push(
-        `Insufficient FA Exception balance (need $${salary}, have $${bucket.remaining}).`
-      );
-      return;
-    }
-    allocateFaExceptionToIncoming({
-      teamCtx: ctx,
-      incomingPlayerId: player.player_id || player.id,
-      amount: salary,
-      bucketType,
-    });
-  });
-
-  if (
-    !violations.length &&
-    faPlayers.some((p) => p.absorptionMode === 'FA_EXCEPTION')
-  ) {
-    if (team.projectedSalary > ctx.context.capSettings?.firstApron) {
-      violations.push('FA Exception usage would exceed First Apron.');
-    } else {
-      markHardCapTriggered(ctx.teamSeasonState, {
-        reason: 'FA_EXCEPTION',
-        season: ctx.context.yearKey,
-      });
-      team.hardCapped = true;
-      team.notes = team.notes || [];
-      faPlayers.forEach((p) => {
-        if (p.absorptionMode === 'FA_EXCEPTION') {
-          const amt = p.matchIncoming ?? getSalaryForYear(p, yearKey);
-          team.notes.push(
-            `Absorbed ${p.name} via ${p.bucketType} bucket for $${amt}; team hard-capped at First Apron.`
-          );
-        }
-      });
-      if (debug.enabled) {
-        debug.faException = summarizeFaExceptionUsage(ctx);
-      }
-    }
-  }
-
-  return violations;
-}
-
 // DRAFT PICKS VALIDATION
-export function validateDraftPicks(team, allTeams) {
-  const violations = [];
-  const currentYear = new Date().getFullYear();
-
-  const unprotectedYears = (team.tradedPicks || [])
-    .filter(
-      (p) =>
-        (p.round === 1 || p.round === '1st') &&
-        !p.isSwap &&
-        !isMeaningfulProtection(p.protection) &&
-        !p.via
-    )
-    .map((p) => parseInt(p.year, 10))
-    .sort((a, b) => a - b);
-
-  for (let i = 1; i < unprotectedYears.length; i++) {
-    if (unprotectedYears[i] === unprotectedYears[i - 1] + 1) {
-      violations.push(
-        `Cannot trade ${unprotectedYears[i - 1]} and ${unprotectedYears[i]} 1st-round picks`
-      );
-    }
-  }
-
-  const limit = currentYear + CBA_THRESHOLDS.STEPIEN_YEARS;
-  const distantPicks = (team.tradedPicks || []).filter((p) => p.year > limit);
-  if (distantPicks.length > 0) {
-    violations.push(`Cannot trade picks beyond ${limit} (7 years out)`);
-  }
-
-  return violations;
-}
-
 // CASH CONSIDERATIONS VALIDATION
-export function validateCash(team, ctx = {}) {
-  const violations = [];
-  const season = ctx.season;
-  const tradesHistory = ctx.tradesHistory || [];
-  const caps = getSeasonalCashCaps(season);
-  const ledger = computeSeasonCashLedger(team.teamId, season, tradesHistory);
-  const projectedSent = (ledger.sentToDate || 0) + (team.cashSent || 0);
-  const projectedReceived =
-    (ledger.receivedToDate || 0) + (team.cashReceived || 0);
-
-  if (
-    team.postTradeStatus?.isAtOrAboveSecondApron &&
-    (team.cashSent > 0 || team.cashReceived > 0)
-  ) {
-    violations.push('Second apron team cannot include cash in trades');
-  }
-
-  if (validationFlags.seasonalCash !== 'off') {
-    if (projectedSent > caps.maxCashOut) {
-      violations.push(
-        `Cash sent exceeds seasonal cap (${formatCurrency(projectedSent)}/${formatCurrency(caps.maxCashOut)})`
-      );
-    }
-    if (projectedReceived > caps.maxCashIn) {
-      violations.push(
-        `Cash received exceeds seasonal cap (${formatCurrency(projectedReceived)}/${formatCurrency(caps.maxCashIn)})`
-      );
-    }
-  }
-
-  return {
-    passed: violations.length === 0,
-    violations,
-    message: violations.length ? 'Cash invalid' : 'Cash valid',
-    details: violations.join('; '),
-  };
-}
-
 // SIGN-AND-TRADE RULES VALIDATION
-export function validateSignAndTrade(team, tradeCtx = {}) {
-  const violations = [];
-  const sntIn = team.incomingPlayers.filter(
-    (p) => p.isSignAndTrade || p.signAndTrade
-  );
-  const sntOut = team.sends.filter((p) => p.signAndTrade);
-  const anySnt = sntIn.length > 0 || sntOut.length > 0;
-
-  if (!anySnt) return violations;
-
-  if (tradeCtx.offseason === false) {
-    violations.push('Sign-and-trade only permitted in offseason');
-  }
-
-  if (sntOut.length > 0) {
-    if (sntOut.length > 0 && (team.sends.length > 1 || team.picksOut.length)) {
-      violations.push('Sign-and-trade player must be traded alone');
-    }
-    sntOut.forEach((p) => {
-      const teamId = team.team?.id ?? team.teamId;
-      if (p.originTeamId && p.originTeamId !== teamId) {
-        violations.push(
-          "Sign-and-trade must be executed by player's original team"
-        );
-      }
-    });
-  }
-
-  if (sntIn.length > 0) {
-    if (
-      team.usedTaxpayerMLEThisSeason ||
-      team.team?.usedTaxpayerMLEThisSeason
-    ) {
-      violations.push(
-        'Teams using taxpayer MLE cannot receive sign-and-trade players'
-      );
-    }
-    if (
-      wouldExceedHardCap(
-        { hardCapTriggered: 'FirstApron' },
-        team.projectedSalary,
-        team.context.capSettings
-      )
-    ) {
-      violations.push('Sign-and-trade would hard-cap team at 1st apron');
-    }
-    team.hardCapped = true;
-    if (team.team) team.team.hardCapTriggered = true;
-    sntIn.forEach((p) => {
-      const years = p.contractYears ?? p.years ?? 0;
-      if (years < 3 || years > 4) {
-        violations.push(`S&T contract for ${p.name} must be 3-4 years`);
-      }
-      if (p.firstYearGuaranteed === false) {
-        violations.push(`First year of ${p.name}\'s deal must be guaranteed`);
-      }
-    });
-  }
-
-  return violations;
-}
-
 // BASE YEAR COMPENSATION VALIDATION
-export function validateBYC(team) {
-  const violations = [];
-
-  // BYC affects salary matching only; no direct contract limit to validate
-
-  return violations;
-}
-
 // SECOND APRON AGGREGATION VALIDATION
-function validateSecondApronRules(team) {
-  const { teamTotalSalary } = team;
-  const { capSettings } = team.context;
-  const { secondApron } = capSettings;
-  const { cashSent = 0, salaryIn, salaryOut } = team;
-  const violations = [];
-
-  if (teamTotalSalary >= secondApron) {
-    if (cashSent > 0) {
-      violations.push('Second apron team cannot include cash in trades');
-    }
-    if (salaryIn > salaryOut) {
-      violations.push('Second apron team cannot receive more salary than sent');
-    }
-  }
-
-  if (!team.overSecondApron && !team.willBeOverSecond) return violations;
-
-  const outgoingSalaries = team.sends
-    .map((p) => getSalaryForYear(p, team.context.yearKey))
-    .sort((a, b) => b - a);
-
-  const incomingSalaries = team.incomingPlayers
-    .map((p) => getSalaryForYear(p, team.context.yearKey))
-    .sort((a, b) => b - a);
-
-  // 1-to-Many rule - Each incoming must be <= single outgoing
-  if (team.sends.length === 1) {
-    const maxOutgoing = outgoingSalaries[0];
-    incomingSalaries.forEach((incoming) => {
-      if (incoming > maxOutgoing) {
-        violations.push(
-          `Second apron restriction: Incoming salary $${incoming.toLocaleString()} ` +
-            `exceeds single outgoing salary $${maxOutgoing.toLocaleString()}`
-        );
-      }
-    });
-  }
-  // Many-to-Many rule - Must pair salaries in descending order
-  else {
-    incomingSalaries.forEach((incoming, i) => {
-      const correspondingOutgoing = outgoingSalaries[i] || 0;
-      if (incoming > correspondingOutgoing) {
-        violations.push(
-          `Second apron salary pairing violation: ` +
-            `Incoming $${incoming.toLocaleString()} > Outgoing $${correspondingOutgoing.toLocaleString()}`
-        );
-      }
-    });
-  }
-
-  // Total salary cannot increase
-  const totalOutgoing = outgoingSalaries.reduce((a, b) => a + b, 0);
-  const totalIncoming = incomingSalaries.reduce((a, b) => a + b, 0);
-  if (totalIncoming > totalOutgoing) {
-    violations.push(
-      `Second apron teams cannot increase total salary: ` +
-        `Incoming $${totalIncoming.toLocaleString()} > Outgoing $${totalOutgoing.toLocaleString()}`
-    );
-  }
-
-  return violations;
-}
-
 // MAIN NEW RULES VALIDATOR
-export function validateAllNewRules(team, allTeams, tradeCtx = {}) {
-  return [
-    ...validateTradeExceptions(team),
-    ...validateDraftPicks(team, allTeams),
-    ...validateCash(team),
-    ...validateSignAndTrade(team, tradeCtx),
-    ...validateBYC(team),
-    ...validateSecondApronRules(team),
-  ];
-}
-function outgoingValueBYC({ newSalary, priorSalary }) {
-  return Math.max(priorSalary, BYC_PERCENT * newSalary);
-}
-
-/** ------------------------------------------------------------------------
- * getMatchingValue(player, year, isOutgoing)
- * A minimal replacement for the legacy helper the tests expect.
- * ▸ For BYC players: outgoing counts max(prior, 50% of new).
- * ▸ For Poison-Pill (rookie max ext.): outgoing uses current salary,
- *   incoming uses average of extension (already baked into salaryByYear).
- * -----------------------------------------------------------------------*/
-const getMatchingValue = (player, yearKey, isOutgoing = false) => {
-  const base = getSalaryForYear(player, yearKey);
-  if (isOutgoing) {
-    if (player.isBYC)
-      return outgoingValueBYC({
-        newSalary: base,
-        priorSalary: player.previousSalary,
-      });
-    if (player.isPoisonPill) return player.currentSalary || base; // PPP
-    return base;
-  }
-
-  if (player.isPoisonPill) {
-    const total =
-      (player.currentSalary ?? base) +
-      (player.extensionYears?.reduce((sum, y) => sum + (y.salary || 0), 0) ||
-        0);
-    const years = 1 + (player.extensionYears?.length || 0);
-    return total / years;
-  }
-
-  return base;
-};
-
-function computeMatchingValues({
-  teams = [],
-  yearKey,
-  daysRemainingInSeason,
-  daysInSeason,
-} = {}) {
-  const proration =
-    Math.min(
-      Math.max((daysRemainingInSeason ?? 0) / (daysInSeason ?? 0), 0),
-      1
-    ) || 1;
-
-  teams.forEach((team) => {
-    (team.sends || []).forEach((player) => {
-      const newSalary = getSalaryForYear(player, yearKey);
-
-      let outgoing = newSalary;
-      if (player.isBYC) {
-        outgoing = outgoingValueBYC({
-          newSalary,
-          priorSalary: player.previousSalary,
-        });
-      } else if (player.isPoisonPill && player.currentSalary) {
-        outgoing = player.currentSalary;
-      }
-      player.matchOutgoing = outgoing;
-
-      let incoming = newSalary;
-      if (player.isPoisonPill) {
-        const total =
-          (player.currentSalary ?? newSalary) +
-          (player.extensionYears?.reduce(
-            (sum, y) => sum + (y.salary || 0),
-            0
-          ) || 0);
-        const years = 1 + (player.extensionYears?.length || 0);
-        incoming = total / years;
-      }
-
-      const pct = Math.min(player.tradeKickerPct ?? 0, 0.15);
-      const waived = Math.min(Math.max(player.tradeKickerWaivedPct ?? 0, 0), 1);
-      const effPct = pct * (1 - waived);
-      if (effPct > 0 && player.remainingGuaranteedOnCurrentContract > 0) {
-        const gross = effPct * player.remainingGuaranteedOnCurrentContract;
-        const currentYearAdd = gross * proration;
-        incoming = (incoming ?? player.currentYearSalary) + currentYearAdd;
-      }
-
-      player.matchIncoming = incoming;
-    });
-  });
-}
-
 const getAllowableIncomingMargin = (team) => {
   const status = team.postTradeStatus || {};
   if (status.isAtOrAboveSecondApron) return 0;
@@ -1907,4 +1050,20 @@ export function hasStepienViolation(picks = []) {
   return false;
 }
 
-export { computeMatchingValues, debug as tradeDebug };
+export {
+  computeMatchingValues,
+  debug as tradeDebug,
+  enforceSecondApronHandcuffs,
+  enforceRosterWindow,
+  enforceConsent,
+  enforceEligibility,
+  enforceTiming,
+  validateTradeExceptions,
+  validateFaExceptionUsage,
+  validateDraftPicks,
+  validateCash,
+  validateSignAndTrade,
+  validateBYC,
+  validateSecondApronRules,
+  validateAllNewRules,
+};
