+++ src/utils/architect/tradeMachine/validators/validateBYC.js
@@ -1,0 +1,7 @@
+export function validateBYC(team) {
+  const violations = [];
+
+  // BYC affects salary matching only; no direct contract limit to validate
+
+  return violations;
+}
