+++ src/utils/architect/tradeMachine/validators/validateDraftPicks.js
@@ -1,0 +1,34 @@
+import { isMeaningfulProtection } from '@/utils/architect/tradeMachine/tradeUtils.js';
+import { CBA_THRESHOLDS } from '@/utils/architect/tradeMachine/cbaConstants.js';
+
+export function validateDraftPicks(team, allTeams) {
+  const violations = [];
+  const currentYear = new Date().getFullYear();
+
+  const unprotectedYears = (team.tradedPicks || [])
+    .filter(
+      (p) =>
+        (p.round === 1 || p.round === '1st') &&
+        !p.isSwap &&
+        !isMeaningfulProtection(p.protection) &&
+        !p.via
+    )
+    .map((p) => parseInt(p.year, 10))
+    .sort((a, b) => a - b);
+
+  for (let i = 1; i < unprotectedYears.length; i++) {
+    if (unprotectedYears[i] === unprotectedYears[i - 1] + 1) {
+      violations.push(
+        `Cannot trade ${unprotectedYears[i - 1]} and ${unprotectedYears[i]} 1st-round picks`
+      );
+    }
+  }
+
+  const limit = currentYear + CBA_THRESHOLDS.STEPIEN_YEARS;
+  const distantPicks = (team.tradedPicks || []).filter((p) => p.year > limit);
+  if (distantPicks.length > 0) {
+    violations.push(`Cannot trade picks beyond ${limit} (7 years out)`);
+  }
+
+  return violations;
+}
